+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - **BREAKING**: Enum `ResultState`.
+- Config `butler.health.heartbeat.report`.
 
 ## [0.5.2] - 2023-12-29
 ### Added